refactor the code to move state to a standalone file
